pyproject.toml
[tool.poetry]
name = "adventure game"
version = "0.1.0"
description = ""
authors = ["Acidicts <111660544+Acidicts@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pygame-ce = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
